refactor(125/WS1.4): extract routes/ (contacts, identity, machines) (#125)
Mechanical extraction of three route groups out of src/server/mod.rs into
a new src/server/routes/ subtree, as the first wave of the #125 server
decomposition:
- routes/contacts.rs  : /contacts* handlers
- routes/identity.rs  : /identity*, /agent* identity block
- routes/machines.rs  : /machines* handlers
- routes/mod.rs       : module wiring + pub(super) re-exports

Pure move: handlers, their colocated #[cfg(test)] modules, and the route
registrations relocate verbatim; visibility widened to pub(super) only
where the extraction requires it. mod.rs shrinks ~1960 lines. The route
set must remain byte-identical (enforced by route_set_matches_registry +
manifest_matches_registry + the tests/snapshots/server-routes.* snapshot).
