+++ src/server/mod.rs
@@ -10,6 +10,7 @@
 use crate as x0x;
 
 mod auth;
+mod routes;
 mod sse;
 mod state;
 mod ws;
@@ -17,6 +18,16 @@ mod ws;
 // Re-export the public server API surface so `x0x::server::*` paths are
 // unchanged after the #125 / WS1.4 extraction. Internal types (AppState,
 // DaemonUpdateConfig, CachedUpgradeCheck) stay private to the crate.
+#[cfg(test)]
+use routes::CardQuery;
+use routes::{
+    add_contact, add_machine, agent_info, agent_sign, agent_user_id_handler, agent_verify,
+    announce_identity, delete_contact, delete_machine, discovered_machine, discovered_machines,
+    get_a2a_agent_card, get_agent_card, import_agent_card, introduction, list_contacts,
+    list_machines, list_revocations, machines_by_user_handler, pin_machine,
+    populate_invite_base_state_from_group_info, quick_trust, revoke_contact, unpin_machine,
+    update_contact,
+};
 use sse::{direct_events_sse, events_sse, peer_events_handler, presence_events, SseEvent};
 pub use state::{
     default_api_address, default_bind_address, default_data_dir, DaemonConfig, ServeOptions,
@@ -46,7 +57,7 @@ use sha2::{Digest, Sha256};
 use std::collections::HashSet;
 use tokio::sync::{broadcast, mpsc, oneshot, watch, Mutex, RwLock};
 use tower_http::cors::CorsLayer;
-use x0x::contacts::{IdentityType, MachineRecord, TrustLevel};
+use x0x::contacts::TrustLevel;
 use x0x::identity::AgentId;
 use x0x::identity::MachineId;
 use x0x::logging::LogHexId;
@@ -287,72 +298,6 @@ struct SubscribeRequest {
     topic: String,
 }
 
-/// POST /agent/sign request body — a caller payload to sign with the
-/// agent's ML-DSA-65 secret key under a mandatory external domain.
-#[derive(Debug, Deserialize)]
-struct AgentSignRequest {
-    /// Required domain-separation string naming the caller's application
-    /// protocol (e.g. `"x0x-symphony-handoff-v1"`). The daemon signs the
-    /// length-prefixed external DST `[0xF0] | magic | len(context) | context |
-    /// payload` (see [`crate::api::agent_signing`]), which is provably disjoint
-    /// from every internal x0x signing input. Must match `[a-z0-9._-]{1,64}`
-    /// and must not name an internal signing domain (issue #133).
-    context: String,
-    /// Base64-encoded bytes to sign. The signature is computed over the DST
-    /// assembled from `context` and these bytes; callers should canonicalize
-    /// structured payloads.
-    payload_b64: String,
-}
-
-/// POST /agent/verify request body — a detached ML-DSA-65 signature to
-/// verify against caller-supplied public key material.
-#[derive(Debug, Deserialize)]
-struct AgentVerifyRequest {
-    /// Base64-encoded bytes the signature was computed over. Same caveat
-    /// as `/agent/sign`: the bytes are taken verbatim, so the caller must
-    /// reproduce the exact canonical serialization that was signed.
-    payload_b64: String,
-    /// Base64-encoded detached ML-DSA-65 signature (3309 bytes decoded).
-    signature_b64: String,
-    /// Base64-encoded ML-DSA-65 public key (1952 bytes decoded).
-    public_key_b64: String,
-    /// Required domain-separation string; verification is performed over
-    /// the same external DST as `/agent/sign`
-    /// (`[0xF0] | magic | len(context) | context | payload`, issue #133), so a
-    /// signature produced by `/agent/sign` round-trips through `/agent/verify`.
-    context: String,
-    /// Optional signing-scheme identifier. When the field is present —
-    /// including as JSON null — it must be exactly
-    /// `x0x.agent-sign.v2.ml-dsa-65`; anything else is rejected with 400
-    /// so a future scheme migration is explicit rather than silent.
-    /// Deserialized via `deserialize_present` because a plain
-    /// `Option<String>` folds present-but-null into `None` and would
-    /// silently accept `"algorithm": null`.
-    #[serde(default, deserialize_with = "deserialize_present")]
-    algorithm: Option<serde_json::Value>,
-}
-
-/// Deserialize a field as `Some(value)` whenever the field is present —
-/// even when the value is JSON null — so present-but-null can be
-/// distinguished from an omitted field (serde's `Option<T>` maps both
-/// to `None`).
-fn deserialize_present<'de, T, D>(deserializer: D) -> Result<Option<T>, D::Error>
-where
-    T: serde::Deserialize<'de>,
-    D: serde::Deserializer<'de>,
-{
-    T::deserialize(deserializer).map(Some)
-}
-
-/// POST /announce request body.
-#[derive(Debug, Default, Deserialize)]
-struct AnnounceIdentityRequest {
-    #[serde(default)]
-    include_user_identity: bool,
-    #[serde(default)]
-    human_consent: bool,
-}
-
 /// POST /task-lists request body.
 #[derive(Debug, Deserialize)]
 struct CreateTaskListRequest {
@@ -372,77 +317,6 @@ struct AddTaskRequest {
 #[derive(Debug, Deserialize)]
 struct UpdateTaskRequest {
     action: String, // "claim" or "complete"
-}
-
-/// POST /contacts request body.
-#[derive(Debug, Deserialize)]
-#[serde(deny_unknown_fields)]
-struct AddContactRequest {
-    /// Agent ID as 64-character hex string.
-    agent_id: String,
-    /// Trust level: "blocked", "unknown", "known", or "trusted".
-    /// Defaults to "known" when omitted.
-    #[serde(default = "default_trust_level")]
-    trust_level: String,
-    /// Optional human-readable label.
-    label: Option<String>,
-}
-
-fn default_trust_level() -> String {
-    "known".to_string()
-}
-
-/// PATCH /contacts/:agent_id request body.
-#[derive(Debug, Deserialize)]
-#[serde(deny_unknown_fields)]
-struct UpdateContactRequest {
-    /// New trust level: "blocked", "unknown", "known", or "trusted".
-    trust_level: Option<String>,
-    /// New identity type: "anonymous", "known", "trusted", or "pinned".
-    identity_type: Option<String>,
-}
-
-/// POST /contacts/:agent_id/machines request body.
-#[derive(Debug, Deserialize)]
-#[serde(deny_unknown_fields)]
-struct AddMachineRequest {
-    /// Machine ID as 64-character hex string.
-    machine_id: String,
-    /// Optional human-readable label.
-    label: Option<String>,
-    /// Whether to pin this machine immediately.
-    #[serde(default)]
-    pinned: bool,
-}
-
-/// Machine record entry for API responses.
-#[derive(Debug, Serialize)]
-struct MachineEntry {
-    machine_id: String,
-    label: Option<String>,
-    first_seen: u64,
-    last_seen: u64,
-    pinned: bool,
-}
-
-/// POST /contacts/trust request body (quick trust shorthand).
-#[derive(Debug, Deserialize)]
-#[serde(deny_unknown_fields)]
-struct QuickTrustRequest {
-    /// Agent ID as 64-character hex string.
-    agent_id: String,
-    /// Trust level: "blocked", "unknown", "known", or "trusted".
-    level: String,
-}
-
-/// Contact entry for API responses.
-#[derive(Debug, Serialize)]
-struct ContactEntry {
-    agent_id: String,
-    trust_level: String,
-    label: Option<String>,
-    added_at: u64,
-    last_seen: Option<u64>,
 }
 
 /// Generic JSON response wrapper.
@@ -473,42 +347,6 @@ struct StatusData {
     agent_id: String,
     peers: usize,
     warnings: Vec<String>,
-}
-
-/// Agent identity response.
-#[derive(Debug, Serialize)]
-struct AgentData {
-    agent_id: String,
-    machine_id: String,
-    user_id: Option<String>,
-    /// Base64 of the agent's ML-KEM-768 public key. Used by other daemons to
-    /// seal group-shared-secret envelopes to this agent.
-    kem_public_key_b64: String,
-}
-
-/// Introduction card response (fields vary by trust level).
-#[derive(Debug, Serialize)]
-struct IntroductionCardData {
-    agent_id: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    machine_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    user_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    certificate: Option<String>,
-    display_name: Option<String>,
-    identity_words: String,
-    services: Vec<ServiceEntryData>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    signature: Option<String>,
-}
-
-/// Service entry in an introduction card.
-#[derive(Debug, Serialize)]
-struct ServiceEntryData {
-    name: String,
-    description: String,
-    min_trust: String,
 }
 
 // ---------------------------------------------------------------------------
@@ -624,13 +462,6 @@ struct MlsDecryptRequest {
     ciphertext: String,
     /// Epoch used for encryption.
     epoch: u64,
-}
-
-/// POST /contacts/:agent_id/revoke request body.
-#[derive(Debug, Deserialize)]
-struct RevokeContactRequest {
-    /// Reason for revocation.
-    reason: String,
 }
 
 /// POST /trust/evaluate request body.
@@ -3316,855 +3147,6 @@ async fn network_status(State(state): State<Arc<AppState>>) -> impl IntoResponse
     )
 }
 
-/// GET /agent
-async fn agent_info(State(state): State<Arc<AppState>>) -> Json<ApiResponse<AgentData>> {
-    use base64::Engine as _;
-    Json(ApiResponse {
-        ok: true,
-        data: AgentData {
-            agent_id: hex::encode(state.agent.agent_id().as_bytes()),
-            machine_id: hex::encode(state.agent.machine_id().as_bytes()),
-            user_id: state.agent.user_id().map(|u| hex::encode(u.as_bytes())),
-            kem_public_key_b64: BASE64.encode(&state.agent_kem_keypair.public_bytes),
-        },
-    })
-}
-
-/// Query parameters for GET /introduction.
-#[derive(Debug, Deserialize)]
-struct IntroductionQuery {
-    /// Connecting peer's agent ID (hex). Determines trust-gated response.
-    #[serde(default)]
-    peer: Option<String>,
-}
-
-/// GET /introduction — serve this agent's introduction card, trust-gated.
-///
-/// Pass `?peer=<hex agent_id>` to receive a card filtered by the peer's
-/// trust level. Without `?peer`, the response is the public (Unknown) view.
-///
-/// - **Blocked**: 403 Forbidden
-/// - **Unknown**: display name, identity words, public services only
-/// - **Known**: above + machine_id, certificate status, broader services
-/// - **Trusted**: everything — all services, full details
-async fn introduction(
-    State(state): State<Arc<AppState>>,
-    axum::extract::Query(query): axum::extract::Query<IntroductionQuery>,
-) -> axum::response::Response {
-    use axum::response::IntoResponse;
-
-    // Resolve the peer's trust level.
-    let peer_trust = if let Some(ref peer_hex) = query.peer {
-        let Ok(peer_bytes) = hex::decode(peer_hex) else {
-            return (
-                StatusCode::BAD_REQUEST,
-                axum::Json(serde_json::json!({"error": "invalid peer agent_id hex"})),
-            )
-                .into_response();
-        };
-        if peer_bytes.len() != 32 {
-            return (
-                StatusCode::BAD_REQUEST,
-                axum::Json(serde_json::json!({"error": "peer agent_id must be 32 bytes"})),
-            )
-                .into_response();
-        }
-        let mut id_bytes = [0u8; 32];
-        id_bytes.copy_from_slice(&peer_bytes);
-        let peer_id = x0x::identity::AgentId(id_bytes);
-        state.contacts.read().await.trust_level(&peer_id)
-    } else {
-        x0x::contacts::TrustLevel::Unknown
-    };
-
-    // Blocked peers get nothing.
-    if peer_trust == x0x::contacts::TrustLevel::Blocked {
-        return (
-            StatusCode::FORBIDDEN,
-            axum::Json(serde_json::json!({"error": "blocked"})),
-        )
-            .into_response();
-    }
-
-    let identity = state.agent.identity();
-
-    // Full service catalogue — filtered below by peer trust.
-    let all_services = vec![
-        x0x::identity::ServiceEntry {
-            name: "presence".to_string(),
-            description: "Online/offline presence visibility".to_string(),
-            min_trust: "unknown".to_string(),
-        },
-        x0x::identity::ServiceEntry {
-            name: "direct-message".to_string(),
-            description: "Send and receive direct encrypted messages".to_string(),
-            min_trust: "known".to_string(),
-        },
-        x0x::identity::ServiceEntry {
-            name: "mls-group".to_string(),
-            description: "Join MLS encrypted group conversations".to_string(),
-            min_trust: "known".to_string(),
-        },
-        x0x::identity::ServiceEntry {
-            name: "file-transfer".to_string(),
-            description: "Send and receive files".to_string(),
-            min_trust: "trusted".to_string(),
-        },
-        x0x::identity::ServiceEntry {
-            name: "payment".to_string(),
-            description: "Payment address exchange".to_string(),
-            min_trust: "trusted".to_string(),
-        },
-    ];
-
-    // Filter services: only return those where peer trust >= min_trust.
-    let peer_rank = peer_trust.rank();
-    let visible_services: Vec<_> = all_services
-        .into_iter()
-        .filter(|s| {
-            s.min_trust
-                .parse::<x0x::contacts::TrustLevel>()
-                .map(|t| peer_rank >= t.rank())
-                .unwrap_or(false)
-        })
-        .collect();
-
-    let card =
-        match x0x::identity::IntroductionCard::from_identity(identity, None, visible_services) {
-            Ok(c) => c,
-            Err(e) => {
-                tracing::warn!("failed to build introduction card: {e}");
-                return (
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    axum::Json(serde_json::json!({
-                        "error": "failed to build introduction card",
-                        "detail": format!("{e}"),
-                    })),
-                )
-                    .into_response();
-            }
-        };
-
-    // Build response — Unknown gets a minimal card, Known/Trusted get progressively more.
-    let data = match peer_trust {
-        x0x::contacts::TrustLevel::Unknown => IntroductionCardData {
-            agent_id: hex::encode(card.agent_id.as_bytes()),
-            machine_id: None,
-            user_id: None,
-            certificate: None,
-            display_name: card.display_name,
-            identity_words: card.identity_words,
-            services: card
-                .services
-                .iter()
-                .map(|s| ServiceEntryData {
-                    name: s.name.clone(),
-                    description: s.description.clone(),
-                    min_trust: s.min_trust.clone(),
-                })
-                .collect(),
-            signature: None,
-        },
-        x0x::contacts::TrustLevel::Known => IntroductionCardData {
-            agent_id: hex::encode(card.agent_id.as_bytes()),
-            machine_id: Some(hex::encode(card.machine_id.as_bytes())),
-            user_id: card.user_id.map(|u| hex::encode(u.as_bytes())),
-            certificate: card.certificate.as_ref().map(|_| "(present)".to_string()),
-            display_name: card.display_name,
-            identity_words: card.identity_words,
-            services: card
-                .services
-                .iter()
-                .map(|s| ServiceEntryData {
-                    name: s.name.clone(),
-                    description: s.description.clone(),
-                    min_trust: s.min_trust.clone(),
-                })
-                .collect(),
-            signature: Some(hex::encode(&card.signature[..8])),
-        },
-        // Trusted — full card.
-        _ => IntroductionCardData {
-            agent_id: hex::encode(card.agent_id.as_bytes()),
-            machine_id: Some(hex::encode(card.machine_id.as_bytes())),
-            user_id: card.user_id.map(|u| hex::encode(u.as_bytes())),
-            certificate: card.certificate.as_ref().map(|_| "(present)".to_string()),
-            display_name: card.display_name,
-            identity_words: card.identity_words,
-            services: card
-                .services
-                .iter()
-                .map(|s| ServiceEntryData {
-                    name: s.name.clone(),
-                    description: s.description.clone(),
-                    min_trust: s.min_trust.clone(),
-                })
-                .collect(),
-            signature: Some(hex::encode(&card.signature[..8])),
-        },
-    };
-
-    axum::Json(ApiResponse { ok: true, data }).into_response()
-}
-
-/// POST /announce — accepts optional JSON body (empty body defaults to no user identity).
-async fn announce_identity(
-    State(state): State<Arc<AppState>>,
-    headers: HeaderMap,
-    body: Bytes,
-) -> impl IntoResponse {
-    let req: AnnounceIdentityRequest = match parse_optional_json(&headers, &body) {
-        Ok(r) => r,
-        Err(resp) => return resp.into_response(),
-    };
-    match state
-        .agent
-        .announce_identity(req.include_user_identity, req.human_consent)
-        .await
-    {
-        Ok(()) => (
-            StatusCode::OK,
-            Json(serde_json::json!({
-                "ok": true,
-                "include_user_identity": req.include_user_identity,
-            })),
-        )
-            .into_response(),
-        Err(e) => bad_request(format!("{e}")).into_response(),
-    }
-}
-
-/// Request body for POST /agent/card/import.
-#[derive(Debug, Deserialize)]
-struct ImportCardRequest {
-    /// Card link (`x0x://agent/...`) or raw base64.
-    card: String,
-    /// Trust level to assign (default: "known").
-    #[serde(default = "default_import_trust")]
-    trust_level: String,
-}
-
-fn default_import_trust() -> String {
-    "known".to_string()
-}
-
-/// Request body for GET /agent/card query params.
-#[derive(Debug, Deserialize)]
-struct CardQuery {
-    /// Display name to include in the card.
-    #[serde(default)]
-    display_name: Option<String>,
-    /// Whether to include group invites.
-    #[serde(default)]
-    include_groups: Option<bool>,
-    /// Include loopback/private interface addresses for local testnet cards.
-    ///
-    /// The default remains false so copy-pasteable cards do not leak
-    /// unroutable RFC1918/loopback addresses to remote recipients.
-    #[serde(default)]
-    include_local_addresses: bool,
-}
-
-/// Populate `addresses` with locally-discovered globally-routable interfaces.
-///
-/// Agent cards are copy-pasteable identity links (`x0x://agent/...`) that can
-/// be shared anywhere. They must only carry globally-advertisable addresses —
-/// a card minted inside a Vultr VPC must not embed `10.200.0.1:5483` or
-/// recipients in London will spend ~50s dialing a black hole.
-fn discover_local_card_addresses(port: u16, addresses: &mut Vec<String>, include_local: bool) {
-    for addr in x0x::collect_local_interface_addrs(port) {
-        if !include_local && !x0x::is_publicly_advertisable(addr) {
-            continue;
-        }
-        let s = addr.to_string();
-        if !addresses.contains(&s) {
-            addresses.push(s);
-        }
-    }
-}
-
-fn prioritize_local_card_addresses(addresses: &mut [String]) {
-    addresses.sort_by_key(|addr| {
-        addr.parse::<std::net::SocketAddr>()
-            .map(x0x::is_publicly_advertisable)
-            .unwrap_or(true)
-    });
-}
-
-fn populate_invite_base_state_from_group_info(
-    invite: &mut x0x::groups::invite::SignedInvite,
-    info: &x0x::groups::GroupInfo,
-) {
-    invite.stable_group_id = Some(info.stable_group_id().to_string());
-    invite.group_created_at = Some(info.created_at);
-    invite.group_description = Some(info.description.clone());
-    invite.policy = Some(info.policy.clone());
-    invite.genesis_creation_nonce = info.genesis.as_ref().map(|g| g.creation_nonce.clone());
-    invite.base_state_revision = Some(info.state_revision);
-    invite.base_state_hash = Some(info.state_hash.clone());
-    invite.base_members_v2 = Some(info.members_v2.clone());
-    invite.base_prev_state_hash = info.prev_state_hash.clone();
-    invite.secure_plane = Some(info.secure_plane);
-    invite.base_secret_epoch = Some(info.secret_epoch);
-    invite.base_security_binding = info.security_binding.clone();
-}
-
-/// GET /agent/card — generate a shareable identity card.
-async fn get_agent_card(
-    State(state): State<Arc<AppState>>,
-    axum::extract::Query(query): axum::extract::Query<CardQuery>,
-) -> impl IntoResponse {
-    let agent_id = state.agent.agent_id();
-    let machine_id = hex::encode(state.agent.machine_id().as_bytes());
-    let display_name = query.display_name.unwrap_or_default();
-
-    let mut card = x0x::groups::card::AgentCard::new(display_name, &agent_id, &machine_id);
-    card.dm_capabilities = Some(x0x::dm::DmCapabilities::v1_gossip_ready(
-        state.agent_kem_keypair.public_bytes.clone(),
-    ));
-
-    // Add user ID if available
-    card.user_id = state.agent.user_id().map(|u| hex::encode(u.as_bytes()));
-
-    // Add external addresses from ant-quic NodeStatus, filtered to
-    // globally-advertisable scope only (see discover_local_card_addresses
-    // doc-comment), then augment with local probes so cards remain useful
-    // before the first observed-address frame arrives from another peer.
-    if let Some(network) = state.agent.network() {
-        if let Some(ns) = network.node_status().await {
-            card.addresses = ns
-                .external_addrs
-                .iter()
-                .filter(|a| query.include_local_addresses || x0x::is_publicly_advertisable(**a))
-                .map(|a| a.to_string())
-                .collect();
-            discover_local_card_addresses(
-                ns.local_addr.port(),
-                &mut card.addresses,
-                query.include_local_addresses,
-            );
-            if query.include_local_addresses {
-                prioritize_local_card_addresses(&mut card.addresses);
-            }
-        }
-    }
-
-    // Optionally include group invite links
-    if query.include_groups.unwrap_or(false) {
-        let groups = state.named_groups.read().await;
-        for info in groups.values() {
-            if info.withdrawn
-                || has_withdrawn_same_stable_group_record(
-                    &groups,
-                    &info.mls_group_id,
-                    Some(info.stable_group_id()),
-                )
-            {
-                continue;
-            }
-            let mut invite = x0x::groups::invite::SignedInvite::new(
-                info.mls_group_id.clone(),
-                info.name.clone(),
-                &agent_id,
-                x0x::groups::invite::DEFAULT_EXPIRY_SECS,
-            );
-            populate_invite_base_state_from_group_info(&mut invite, info);
-            card.groups.push(x0x::groups::card::CardGroup {
-                name: info.name.clone(),
-                invite_link: invite.to_link(),
-            });
-        }
-    }
-
-    // Include stores
-    let stores = state.kv_stores.read().await;
-    for (topic, _) in stores.iter() {
-        card.stores.push(x0x::groups::card::CardStore {
-            name: topic.clone(),
-            topic: topic.clone(),
-        });
-    }
-
-    // Sign the card (ADR-0017) so its reachability hints and capability
-    // advertisements are tamper-evident in transit. Signing should not fail
-    // for a valid keypair; degrade to an unsigned card with a warning rather
-    // than failing the request.
-    if let Err(e) = card.sign(state.agent.identity().agent_keypair()) {
-        tracing::warn!("failed to sign agent card: {e}");
-    }
-
-    let link = card.to_link();
-
-    Json(serde_json::json!({
-        "ok": true,
-        "card": card,
-        "link": link,
-    }))
-}
-
-/// GET /.well-known/agent-card.json — A2A-compatible discovery card (ADR-0017).
-///
-/// Serves the local agent's identity as a Google A2A Agent Card so the agent
-/// is discoverable by the A2A ecosystem. The underlying x0x card is signed,
-/// and the signature/public key are surfaced as `x0x`-namespaced extensions.
-async fn get_a2a_agent_card(State(state): State<Arc<AppState>>) -> impl IntoResponse {
-    let agent_id = state.agent.agent_id();
-    let machine_id = hex::encode(state.agent.machine_id().as_bytes());
-
-    let mut card = x0x::groups::card::AgentCard::new(String::new(), &agent_id, &machine_id);
-    card.dm_capabilities = Some(x0x::dm::DmCapabilities::v1_gossip_ready(
-        state.agent_kem_keypair.public_bytes.clone(),
-    ));
-    card.user_id = state.agent.user_id().map(|u| hex::encode(u.as_bytes()));
-
-    // Only globally-advertisable addresses belong in a publicly-served card.
-    if let Some(network) = state.agent.network() {
-        if let Some(ns) = network.node_status().await {
-            card.addresses = ns
-                .external_addrs
-                .iter()
-                .filter(|a| x0x::is_publicly_advertisable(**a))
-                .map(|a| a.to_string())
-                .collect();
-        }
-    }
-
-    // Public stores become A2A skills.
-    {
-        let stores = state.kv_stores.read().await;
-        for (topic, _) in stores.iter() {
-            card.stores.push(x0x::groups::card::CardStore {
-                name: topic.clone(),
-                topic: topic.clone(),
-            });
-        }
-    }
-
-    if let Err(e) = card.sign(state.agent.identity().agent_keypair()) {
-        tracing::warn!("failed to sign A2A agent card: {e}");
-    }
-
-    let certificate_b64 = state.agent.identity().agent_certificate().and_then(|c| {
-        use base64::Engine;
-        bincode::serialize(c)
-            .ok()
-            .map(|b| base64::engine::general_purpose::STANDARD.encode(b))
-    });
-
-    let ctx = x0x::a2a::A2aContext {
-        version: env!("CARGO_PKG_VERSION").to_string(),
-        exec_enabled: state.exec_service.enabled(),
-        certificate_b64,
-    };
-
-    // `Json` sets `content-type: application/json`.
-    Json(x0x::a2a::a2a_card_from(&card, &ctx))
-}
-
-/// POST /agent/card/import — import an agent card to contacts.
-async fn import_agent_card(
-    State(state): State<Arc<AppState>>,
-    Json(req): Json<ImportCardRequest>,
-) -> impl IntoResponse {
-    // Parse card
-    let card = match x0x::groups::card::AgentCard::from_link(&req.card) {
-        Ok(c) => c,
-        Err(e) => {
-            return bad_request(format!("invalid card: {e}"));
-        }
-    };
-
-    // ADR-0017: reject tampered signed cards. A signed card whose signature
-    // fails verification (or whose embedded key does not match its agent_id)
-    // is dropped. Legacy unsigned cards (signature == None) remain importable
-    // for backward compatibility.
-    if card.signature.is_some() {
-        if let Err(e) = card.verify_signature() {
-            return bad_request(format!("agent card signature invalid: {e}"));
-        }
-    }
-
-    // Parse trust level — surface the FromStr error rather than silently
-    // coercing unknown values to Known. Matches the AddContactRequest path.
-    let trust: x0x::contacts::TrustLevel = match req.trust_level.parse() {
-        Ok(t) => t,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(serde_json::json!({ "ok": false, "error": e })),
-            );
-        }
-    };
-
-    // Parse agent ID
-    let agent_id_bytes: [u8; 32] = match hex::decode(&card.agent_id) {
-        Ok(bytes) if bytes.len() == 32 => {
-            let mut arr = [0u8; 32];
-            arr.copy_from_slice(&bytes);
-            arr
-        }
-        _ => {
-            return bad_request("invalid agent_id in card");
-        }
-    };
-    let agent_id = x0x::identity::AgentId(agent_id_bytes);
-
-    // Add to contacts
-    let now = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|d| d.as_secs())
-        .unwrap_or(0);
-
-    let contact = x0x::contacts::Contact {
-        agent_id,
-        trust_level: trust,
-        label: Some(card.display_name.clone()),
-        added_at: now,
-        last_seen: None,
-        identity_type: x0x::contacts::IdentityType::default(),
-        machines: Vec::new(),
-        dm_capabilities: card.dm_capabilities.clone(),
-    };
-
-    state.contacts.write().await.add(contact);
-
-    // Also populate the identity discovery cache so connect_to_agent / send_direct
-    // can find this agent without waiting for gossip announcements.
-    let machine_id_bytes: [u8; 32] = hex::decode(&card.machine_id)
-        .ok()
-        .and_then(|b| b.try_into().ok())
-        .unwrap_or([0u8; 32]);
-    let addresses: Vec<std::net::SocketAddr> = card
-        .addresses
-        .iter()
-        .filter_map(|a| a.parse().ok())
-        .collect();
-
-    let capability_store = state.agent.capability_store();
-    let mut inserted_dm_capability = false;
-    if machine_id_bytes != [0u8; 32] {
-        if let Some(caps) = card.dm_capabilities.clone() {
-            if caps.gossip_inbox && !caps.kem_public_key.is_empty() {
-                capability_store.insert(
-                    agent_id,
-                    x0x::identity::MachineId(machine_id_bytes),
-                    caps,
-                    x0x::dm_capability::now_unix_ms(),
-                );
-                inserted_dm_capability = true;
-            }
-        }
-    }
-    tracing::debug!(
-        target: "dm.trace",
-        stage = "agent_card_import_capability",
-        agent_id = %hex::encode(agent_id.as_bytes()),
-        machine_id = %hex::encode(machine_id_bytes),
-        card_has_capability = card.dm_capabilities.is_some(),
-        inserted = inserted_dm_capability,
-        capability_store_entries = capability_store.len(),
-    );
-
-    if machine_id_bytes != [0u8; 32] || !addresses.is_empty() {
-        state
-            .agent
-            .insert_discovered_agent_for_testing(x0x::DiscoveredAgent {
-                agent_id,
-                machine_id: x0x::identity::MachineId(machine_id_bytes),
-                user_id: None,
-                addresses,
-                announced_at: now,
-                last_seen: now,
-                machine_public_key: Vec::new(),
-                nat_type: None,
-                can_receive_direct: None,
-                is_relay: None,
-                is_coordinator: None,
-                reachable_via: Vec::new(),
-                relay_candidates: Vec::new(),
-            })
-            .await;
-    }
-
-    (
-        StatusCode::OK,
-        Json(serde_json::json!({
-            "ok": true,
-            "agent_id": card.agent_id,
-            "display_name": card.display_name,
-            "trust_level": format!("{trust:?}"),
-            "groups": card.groups.len(),
-            "stores": card.stores.len(),
-        })),
-    )
-}
-
-/// Maximum payload size accepted by `POST /agent/sign` / `/agent/verify`,
-/// in bytes. External signing is for hashes, manifests, and audit records —
-/// not blobs. Mirrors [`crate::api::agent_signing::MAX_PAYLOAD_BYTES`] (kept
-/// as a local const so the 413 path can format the limit without importing
-/// the helper into every handler call site).
-const AGENT_SIGN_MAX_PAYLOAD_BYTES: usize = crate::api::agent_signing::MAX_PAYLOAD_BYTES;
-
-/// Stable scheme identifier returned by `POST /agent/sign` and accepted by
-/// `/agent/verify`. The `v2` scheme signs the domain-separated external DST
-/// (issue #133); the pre-#133 `v1` scheme (optional `domain || 0x00 ||
-/// payload` / raw payload) is no longer produced.
-const AGENT_SIGN_SCHEME_ID: &str = crate::api::agent_signing::SCHEME_ID;
-
-/// POST /agent/sign — produce a detached ML-DSA-65 signature over a
-/// caller-supplied payload using this agent's signing key.
-///
-/// Rationale. x0xd already signs gossip frames at the transport layer
-/// (saorsa-gossip-identity), but transport-layer signatures don't survive
-/// a database read. Applications that persist signed records to disk or
-/// to distributed storage (audit logs, governance votes, content
-/// metadata) need a detached signature that can be verified later from
-/// the stored bytes alone, by a verifier that may have never been on the
-/// network when the signature was issued. This endpoint provides that
-/// primitive without exposing the secret key itself.
-///
-/// Authentication. Bearer-token authenticated like every other endpoint
-/// — only callers with the agent's local API token can sign as the agent.
-///
-/// Payload. `payload_b64` is base64-decoded to the raw payload, which is
-/// taken verbatim (the caller owns the canonical serialization of any
-/// structured payload — e.g. `serde_canonical_json`, `postcard`, or an
-/// explicit field-order convention). Payloads are capped at 64 KiB:
-/// external signing is for hashes, manifests, and audit records, not blobs.
-///
-/// Domain separation (issue #133, mandatory). The signature is *never*
-/// computed over the raw payload. A required `context` string — matching
-/// `[a-z0-9._-]{1,64}` and not naming an internal x0x signing domain
-/// (see `INTERNAL_CONTEXT_DENYLIST` in `src/api/agent_signing.rs`) — binds
-/// the signature to a single application protocol. The canonical signed
-/// bytes are the external DST
-///
-/// ```text
-/// [0xF0] | b"x0x.external-agent-sign.v1" | len(context):u32 BE | context | payload
-/// ```
-///
-/// That prologue is provably disjoint from every internal x0x signing
-/// input (none begins with `[0xF0] | magic`), so an external signature
-/// can never be replayed as a protocol message; the `0xF0` namespace tag
-/// and the length-prefixed context make the boundary unambiguous. The
-/// `context` is echoed in the response so a verifier knows the
-/// canonical-bytes shape without out-of-band information.
-///
-/// Scheme. Returns the stable identifier `x0x.agent-sign.v2.ml-dsa-65`.
-/// The `.v2` pins the API-envelope version; the magic's `.v1` pins the DST
-/// byte layout — two independent axes (see `src/api/agent_signing.rs`). A
-/// future scheme migration is therefore explicit in the response, not
-/// silent.
-///
-/// Response. Returns the agent_id (hex, 32 bytes), the agent's public
-/// key (base64), the signature (base64), the context (echoed), and the
-/// scheme identifier. All values are wire-format ready for inclusion in
-/// the signed record.
-async fn agent_sign(
-    State(state): State<Arc<AppState>>,
-    Json(req): Json<AgentSignRequest>,
-) -> impl IntoResponse {
-    let payload = match BASE64.decode(&req.payload_b64) {
-        Ok(bytes) => bytes,
-        Err(e) => {
-            return bad_request(format!("invalid base64 payload: {e}"));
-        }
-    };
-
-    if payload.is_empty() {
-        return bad_request("payload must be non-empty");
-    }
-
-    if payload.len() > AGENT_SIGN_MAX_PAYLOAD_BYTES {
-        return api_error(
-            StatusCode::PAYLOAD_TOO_LARGE,
-            format!(
-                "payload exceeds maximum signable size of {} bytes",
-                AGENT_SIGN_MAX_PAYLOAD_BYTES
-            ),
-        );
-    }
-
-    // Mandatory external domain separation (issue #133): sign the
-    // length-prefixed external DST `[0xF0] | magic | len(context) | context |
-    // payload`, which is provably disjoint from every internal x0x signing
-    // input (see `src/api/agent_signing.rs`). `context` is required and
-    // validated — there is no raw-payload signing path.
-    if let Err(e) = crate::api::agent_signing::validate_context(&req.context) {
-        return bad_request(e.to_string());
-    }
-    let canonical = crate::api::agent_signing::assemble_buffer(&req.context, &payload);
-
-    let identity = state.agent.identity();
-    let keypair = identity.agent_keypair();
-
-    let signature = match ant_quic::crypto::raw_public_keys::pqc::sign_with_ml_dsa(
-        keypair.secret_key(),
-        &canonical,
-    ) {
-        Ok(sig) => sig,
-        Err(e) => {
-            return api_error(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                format!("signing failed: {e:?}"),
-            );
-        }
-    };
-
-    let signature_b64 = BASE64.encode(signature.as_bytes());
-    let public_key_b64 = BASE64.encode(keypair.public_key().as_bytes());
-    let agent_id_hex = hex::encode(state.agent.agent_id().as_bytes());
-
-    let mut resp = serde_json::json!({
-        "ok": true,
-        "agent_id": agent_id_hex,
-        "public_key_b64": public_key_b64,
-        "signature_b64": signature_b64,
-        "algorithm": AGENT_SIGN_SCHEME_ID,
-    });
-    // Echo the context so a verifier knows the canonical-bytes shape
-    // without out-of-band context.
-    resp["context"] = serde_json::Value::String(req.context);
-
-    (StatusCode::OK, Json(resp))
-}
-
-/// POST /agent/verify — verify a detached ML-DSA-65 signature against a
-/// caller-supplied public key (issue #106).
-///
-/// Rationale. The counterpart to `POST /agent/sign`: applications that
-/// persist signed records read them back — often on machines that never
-/// authored them — and must verify from the stored bytes alone. Without
-/// this endpoint every consumer would bundle its own FIPS-204 library and
-/// re-derive x0x's canonical external DST framing, which would drift the
-/// moment the convention evolves.
-///
-/// Statelessness. Verification uses only caller-supplied public material:
-/// no key access, no identity state. The handler deliberately takes no
-/// `State` extractor so this property is enforced at compile time.
-///
-/// Authentication. Bearer-token authenticated like every other endpoint.
-///
-/// Semantics. A failed signature check is a *result*, not an error:
-/// `200` with `valid: false`. `400` is reserved for malformed input (bad
-/// base64, empty payload, wrong key or signature length, an invalid or
-/// internal-reserved `context`, or an unknown `algorithm`); `413` for
-/// payloads over the 64 KiB cap — mirroring `/agent/sign` exactly.
-/// Verification is performed over the *same* external DST as signing,
-/// `[0xF0] | magic | len(context):u32 BE | context | payload`, using the
-/// caller-supplied `context` (required, validated identically to
-/// `/agent/sign`). A signature produced for one context therefore does
-/// not verify under any other — and raw-payload verification is no longer
-/// a valid input.
-async fn agent_verify(Json(req): Json<AgentVerifyRequest>) -> impl IntoResponse {
-    let payload = match BASE64.decode(&req.payload_b64) {
-        Ok(bytes) => bytes,
-        Err(e) => {
-            return bad_request(format!("invalid base64 payload: {e}"));
-        }
-    };
-
-    if payload.is_empty() {
-        return bad_request("payload must be non-empty");
-    }
-
-    if payload.len() > AGENT_SIGN_MAX_PAYLOAD_BYTES {
-        return api_error(
-            StatusCode::PAYLOAD_TOO_LARGE,
-            format!(
-                "payload exceeds maximum verifiable size of {} bytes",
-                AGENT_SIGN_MAX_PAYLOAD_BYTES
-            ),
-        );
-    }
-
-    // Same canonical-bytes assembly as `agent_sign`: the verifier must
-    // reconstruct exactly the bytes the signer committed to.
-    if let Err(e) = crate::api::agent_signing::validate_context(&req.context) {
-        return bad_request(e.to_string());
-    }
-    let canonical = crate::api::agent_signing::assemble_buffer(&req.context, &payload);
-
-    // A present `algorithm` must be exactly the supported scheme string;
-    // JSON null and non-string values are present-but-wrong, not omitted.
-    if let Some(algorithm) = req.algorithm.as_ref() {
-        if algorithm.as_str() != Some(AGENT_SIGN_SCHEME_ID) {
-            return bad_request(format!(
-                "unsupported algorithm: {algorithm} (expected {AGENT_SIGN_SCHEME_ID})"
-            ));
-        }
-    }
-
-    let signature_bytes = match BASE64.decode(&req.signature_b64) {
-        Ok(bytes) => bytes,
-        Err(e) => {
-            return bad_request(format!("invalid base64 signature: {e}"));
-        }
-    };
-
-    // A wrong-length signature is malformed input, not a failed check —
-    // reject it with 400 so a truncated paste never reads as `valid: false`.
-    if signature_bytes.len() != ant_quic::crypto::raw_public_keys::pqc::ML_DSA_65_SIGNATURE_SIZE {
-        return bad_request(format!(
-            "signature must be exactly {} bytes for ML-DSA-65, got {}",
-            ant_quic::crypto::raw_public_keys::pqc::ML_DSA_65_SIGNATURE_SIZE,
-            signature_bytes.len()
-        ));
-    }
-
-    let signature = match ant_quic::crypto::raw_public_keys::pqc::MlDsaSignature::from_bytes(
-        &signature_bytes,
-    ) {
-        Ok(sig) => sig,
-        Err(e) => {
-            return bad_request(format!("invalid signature format: {e:?}"));
-        }
-    };
-
-    let public_key_bytes = match BASE64.decode(&req.public_key_b64) {
-        Ok(bytes) => bytes,
-        Err(e) => {
-            return bad_request(format!("invalid base64 public key: {e}"));
-        }
-    };
-
-    // An ML-DSA-65 public key is exactly 1952 bytes; anything else is a
-    // wrong-key-type paste and gets 400, never a confusing `valid: false`.
-    if public_key_bytes.len() != ant_quic::crypto::raw_public_keys::pqc::ML_DSA_65_PUBLIC_KEY_SIZE {
-        return bad_request(format!(
-            "public key must be exactly {} bytes for ML-DSA-65, got {}",
-            ant_quic::crypto::raw_public_keys::pqc::ML_DSA_65_PUBLIC_KEY_SIZE,
-            public_key_bytes.len()
-        ));
-    }
-
-    let public_key =
-        match ant_quic::crypto::raw_public_keys::pqc::MlDsaPublicKey::from_bytes(&public_key_bytes)
-        {
-            Ok(pk) => pk,
-            Err(e) => {
-                return bad_request(format!("invalid public key format: {e:?}"));
-            }
-        };
-
-    let valid = ant_quic::crypto::raw_public_keys::pqc::verify_with_ml_dsa(
-        &public_key,
-        &canonical,
-        &signature,
-    )
-    .is_ok();
-
-    (
-        StatusCode::OK,
-        Json(serde_json::json!({
-            "ok": true,
-            "valid": valid,
-            "algorithm": AGENT_SIGN_SCHEME_ID,
-        })),
-    )
-}
-
 /// GET /peers
 async fn peers(State(state): State<Arc<AppState>>) -> impl IntoResponse {
     match state.agent.peers().await {
@@ -4591,78 +3573,6 @@ async fn discovered_agent(
     }
 }
 
-/// GET /machines/discovered
-async fn discovered_machines(
-    State(state): State<Arc<AppState>>,
-    Query(query): Query<DiscoveredAgentsQuery>,
-) -> impl IntoResponse {
-    let result = if query.unfiltered {
-        state.agent.discovered_machines_unfiltered().await
-    } else {
-        state.agent.discovered_machines().await
-    };
-    match result {
-        Ok(machines) => {
-            let entries: Vec<DiscoveredMachineEntry> =
-                machines.into_iter().map(discovered_machine_entry).collect();
-            (
-                StatusCode::OK,
-                Json(serde_json::json!({ "ok": true, "machines": entries })),
-            )
-        }
-        Err(e) => api_error(StatusCode::INTERNAL_SERVER_ERROR, format!("{e}")),
-    }
-}
-
-/// GET /machines/discovered/:machine_id
-async fn discovered_machine(
-    State(state): State<Arc<AppState>>,
-    Path(machine_id_hex): Path<String>,
-    Query(params): Query<DiscoveredAgentQuery>,
-) -> impl IntoResponse {
-    let machine_id = match parse_machine_id_hex(&machine_id_hex) {
-        Ok(id) => id,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(serde_json::json!({ "ok": false, "error": e })),
-            );
-        }
-    };
-
-    if params.wait {
-        match state.agent.find_machine(machine_id, 10).await {
-            Ok(Some(machine)) => {
-                return (
-                    StatusCode::OK,
-                    Json(serde_json::json!({
-                        "ok": true,
-                        "machine": discovered_machine_entry(machine),
-                    })),
-                );
-            }
-            Ok(None) => {
-                return not_found("machine not found within timeout");
-            }
-            Err(e) => {
-                return api_error(StatusCode::INTERNAL_SERVER_ERROR, format!("{e}"));
-            }
-        }
-    }
-
-    match state.agent.discovered_machine(machine_id).await {
-        Ok(Some(machine)) => (
-            StatusCode::OK,
-            Json(serde_json::json!({
-                "ok": true,
-                "machine": discovered_machine_entry(machine),
-            })),
-        ),
-        Ok(None) => not_found("machine not found"),
-        Err(e) => api_error(StatusCode::INTERNAL_SERVER_ERROR, format!("{e}")),
-    }
-}
-
 /// GET /agents/:agent_id/machine
 async fn machine_for_agent_handler(
     State(state): State<Arc<AppState>>,
@@ -4725,48 +3635,6 @@ async fn agents_by_user_handler(
     }
 }
 
-/// GET /users/:user_id/machines
-async fn machines_by_user_handler(
-    State(state): State<Arc<AppState>>,
-    Path(user_id_hex): Path<String>,
-) -> (StatusCode, Json<serde_json::Value>) {
-    let user_id_bytes = match hex::decode(&user_id_hex) {
-        Ok(b) if b.len() == 32 => {
-            let mut arr = [0u8; 32];
-            arr.copy_from_slice(&b);
-            arr
-        }
-        _ => {
-            return bad_request("invalid user_id: expected 64 hex characters");
-        }
-    };
-    let user_id = x0x::identity::UserId(user_id_bytes);
-    match state.agent.find_machines_by_user(user_id).await {
-        Ok(machines) => {
-            let entries: Vec<DiscoveredMachineEntry> =
-                machines.into_iter().map(discovered_machine_entry).collect();
-            (
-                StatusCode::OK,
-                Json(serde_json::json!({
-                    "ok": true,
-                    "user_id": user_id_hex,
-                    "machines": entries,
-                })),
-            )
-        }
-        Err(e) => api_error(StatusCode::INTERNAL_SERVER_ERROR, format!("{e}")),
-    }
-}
-
-/// GET /agent/user-id
-async fn agent_user_id_handler(State(state): State<Arc<AppState>>) -> Json<serde_json::Value> {
-    let user_id = state.agent.user_id().map(|uid| hex::encode(uid.0));
-    Json(serde_json::json!({
-        "ok": true,
-        "user_id": user_id,
-    }))
-}
-
 // ---------------------------------------------------------------------------
 // Contact handlers
 // ---------------------------------------------------------------------------
@@ -4797,319 +3665,6 @@ fn parse_machine_id_hex(hex_str: &str) -> Result<MachineId, String> {
     let mut arr = [0u8; 32];
     arr.copy_from_slice(&bytes);
     Ok(MachineId(arr))
-}
-
-/// GET /contacts — list all contacts with trust levels.
-async fn list_contacts(State(state): State<Arc<AppState>>) -> impl IntoResponse {
-    let store = state.contacts.read().await;
-    let entries: Vec<ContactEntry> = store
-        .list()
-        .into_iter()
-        .map(|c| ContactEntry {
-            agent_id: hex::encode(c.agent_id.0),
-            trust_level: c.trust_level.to_string(),
-            label: c.label.clone(),
-            added_at: c.added_at,
-            last_seen: c.last_seen,
-        })
-        .collect();
-    Json(serde_json::json!({ "ok": true, "contacts": entries }))
-}
-
-/// POST /contacts — add a new contact.
-async fn add_contact(
-    State(state): State<Arc<AppState>>,
-    Json(req): Json<AddContactRequest>,
-) -> impl IntoResponse {
-    let agent_id = match parse_agent_id_hex(&req.agent_id) {
-        Ok(id) => id,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(serde_json::json!({ "ok": false, "error": e })),
-            );
-        }
-    };
-
-    let trust_level: TrustLevel = match req.trust_level.parse() {
-        Ok(t) => t,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(serde_json::json!({ "ok": false, "error": e })),
-            );
-        }
-    };
-
-    let now = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|d| d.as_secs())
-        .unwrap_or(0);
-
-    let contact = x0x::contacts::Contact {
-        agent_id,
-        trust_level,
-        label: req.label,
-        added_at: now,
-        last_seen: None,
-        identity_type: x0x::contacts::IdentityType::default(),
-        machines: Vec::new(),
-        dm_capabilities: None,
-    };
-
-    state.contacts.write().await.add(contact);
-
-    (
-        StatusCode::CREATED,
-        Json(serde_json::json!({ "ok": true, "agent_id": hex::encode(agent_id.0) })),
-    )
-}
-
-/// PATCH /contacts/:agent_id — update trust level and/or identity type for a contact.
-async fn update_contact(
-    State(state): State<Arc<AppState>>,
-    Path(agent_id_hex): Path<String>,
-    Json(req): Json<UpdateContactRequest>,
-) -> impl IntoResponse {
-    let agent_id = match parse_agent_id_hex(&agent_id_hex) {
-        Ok(id) => id,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(serde_json::json!({ "ok": false, "error": e })),
-            );
-        }
-    };
-
-    let mut store = state.contacts.write().await;
-
-    if let Some(ref tl_str) = req.trust_level {
-        let trust_level: TrustLevel = match tl_str.parse() {
-            Ok(t) => t,
-            Err(e) => {
-                return (
-                    StatusCode::BAD_REQUEST,
-                    Json(serde_json::json!({ "ok": false, "error": e })),
-                );
-            }
-        };
-        store.set_trust(&agent_id, trust_level);
-    }
-
-    if let Some(ref it_str) = req.identity_type {
-        let identity_type: IdentityType = match it_str.parse() {
-            Ok(t) => t,
-            Err(e) => {
-                return (
-                    StatusCode::BAD_REQUEST,
-                    Json(serde_json::json!({ "ok": false, "error": e })),
-                );
-            }
-        };
-        store.set_identity_type(&agent_id, identity_type);
-    }
-
-    (StatusCode::OK, Json(serde_json::json!({ "ok": true })))
-}
-
-/// GET /contacts/:agent_id/machines — list machine records for a contact.
-async fn list_machines(
-    State(state): State<Arc<AppState>>,
-    Path(agent_id_hex): Path<String>,
-) -> impl IntoResponse {
-    let agent_id = match parse_agent_id_hex(&agent_id_hex) {
-        Ok(id) => id,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(serde_json::json!({ "ok": false, "error": e })),
-            )
-                .into_response();
-        }
-    };
-
-    let store = state.contacts.read().await;
-    let entries: Vec<MachineEntry> = store
-        .machines(&agent_id)
-        .iter()
-        .map(|m| MachineEntry {
-            machine_id: hex::encode(m.machine_id.0),
-            label: m.label.clone(),
-            first_seen: m.first_seen,
-            last_seen: m.last_seen,
-            pinned: m.pinned,
-        })
-        .collect();
-
-    (
-        StatusCode::OK,
-        Json(serde_json::json!({ "ok": true, "machines": entries })),
-    )
-        .into_response()
-}
-
-/// POST /contacts/:agent_id/machines — add a machine record for a contact.
-async fn add_machine(
-    State(state): State<Arc<AppState>>,
-    Path(agent_id_hex): Path<String>,
-    Json(req): Json<AddMachineRequest>,
-) -> impl IntoResponse {
-    let agent_id = match parse_agent_id_hex(&agent_id_hex) {
-        Ok(id) => id,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(serde_json::json!({ "ok": false, "error": e })),
-            )
-                .into_response();
-        }
-    };
-
-    let machine_bytes = match hex::decode(&req.machine_id) {
-        Ok(b) if b.len() == 32 => {
-            let mut arr = [0u8; 32];
-            arr.copy_from_slice(&b);
-            arr
-        }
-        _ => {
-            return bad_request("machine_id must be a 64-character hex string").into_response();
-        }
-    };
-    let machine_id = MachineId(machine_bytes);
-
-    let record = MachineRecord::new(machine_id, req.label.clone());
-    let mut store = state.contacts.write().await;
-    let is_new = store.add_machine(&agent_id, record);
-
-    if req.pinned {
-        store.pin_machine(&agent_id, &machine_id);
-    }
-
-    let status = if is_new {
-        StatusCode::CREATED
-    } else {
-        StatusCode::OK
-    };
-    let entry = MachineEntry {
-        machine_id: hex::encode(machine_id.0),
-        label: req.label,
-        first_seen: store
-            .machines(&agent_id)
-            .iter()
-            .find(|m| m.machine_id == machine_id)
-            .map(|m| m.first_seen)
-            .unwrap_or(0),
-        last_seen: store
-            .machines(&agent_id)
-            .iter()
-            .find(|m| m.machine_id == machine_id)
-            .map(|m| m.last_seen)
-            .unwrap_or(0),
-        pinned: req.pinned,
-    };
-
-    (
-        status,
-        Json(serde_json::json!({ "ok": true, "machine": entry })),
-    )
-        .into_response()
-}
-
-/// DELETE /contacts/:agent_id/machines/:machine_id — remove a machine record.
-async fn delete_machine(
-    State(state): State<Arc<AppState>>,
-    Path((agent_id_hex, machine_id_hex)): Path<(String, String)>,
-) -> impl IntoResponse {
-    let agent_id = match parse_agent_id_hex(&agent_id_hex) {
-        Ok(id) => id,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(serde_json::json!({ "ok": false, "error": e })),
-            )
-                .into_response();
-        }
-    };
-
-    let machine_bytes = match hex::decode(&machine_id_hex) {
-        Ok(b) if b.len() == 32 => {
-            let mut arr = [0u8; 32];
-            arr.copy_from_slice(&b);
-            arr
-        }
-        _ => {
-            return bad_request("machine_id must be a 64-character hex string").into_response();
-        }
-    };
-    let machine_id = MachineId(machine_bytes);
-
-    let removed = state
-        .contacts
-        .write()
-        .await
-        .remove_machine(&agent_id, &machine_id);
-    if removed {
-        (StatusCode::NO_CONTENT, Json(serde_json::json!({}))).into_response()
-    } else {
-        not_found("machine not found").into_response()
-    }
-}
-
-/// DELETE /contacts/:agent_id — remove a contact.
-async fn delete_contact(
-    State(state): State<Arc<AppState>>,
-    Path(agent_id_hex): Path<String>,
-) -> impl IntoResponse {
-    let agent_id = match parse_agent_id_hex(&agent_id_hex) {
-        Ok(id) => id,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(serde_json::json!({ "ok": false, "error": e })),
-            );
-        }
-    };
-
-    let removed = state.contacts.write().await.remove(&agent_id);
-    if removed.is_some() {
-        (StatusCode::OK, Json(serde_json::json!({ "ok": true })))
-    } else {
-        not_found("contact not found")
-    }
-}
-
-/// POST /contacts/trust — quick trust shorthand.
-async fn quick_trust(
-    State(state): State<Arc<AppState>>,
-    Json(req): Json<QuickTrustRequest>,
-) -> impl IntoResponse {
-    let agent_id = match parse_agent_id_hex(&req.agent_id) {
-        Ok(id) => id,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(serde_json::json!({ "ok": false, "error": e })),
-            );
-        }
-    };
-
-    let trust_level: TrustLevel = match req.level.parse() {
-        Ok(t) => t,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(serde_json::json!({ "ok": false, "error": e })),
-            );
-        }
-    };
-
-    state
-        .contacts
-        .write()
-        .await
-        .set_trust(&agent_id, trust_level);
-
-    (StatusCode::OK, Json(serde_json::json!({ "ok": true })))
 }
 
 // ---------------------------------------------------------------------------
@@ -16128,121 +14683,6 @@ async fn agent_reachability(
 // ---------------------------------------------------------------------------
 // Contact trust extension handlers
 // ---------------------------------------------------------------------------
-
-/// POST /contacts/:agent_id/revoke — permanently revoke an agent's key.
-async fn revoke_contact(
-    State(state): State<Arc<AppState>>,
-    Path(agent_id_hex): Path<String>,
-    Json(req): Json<RevokeContactRequest>,
-) -> impl IntoResponse {
-    let agent_id = match parse_agent_id_hex(&agent_id_hex) {
-        Ok(id) => id,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(serde_json::json!({ "ok": false, "error": e })),
-            );
-        }
-    };
-
-    let mut store = state.contacts.write().await;
-    store.revoke(&agent_id, &req.reason);
-    (StatusCode::OK, Json(serde_json::json!({ "ok": true })))
-}
-
-/// GET /contacts/:agent_id/revocations — list revocation records.
-async fn list_revocations(State(state): State<Arc<AppState>>) -> impl IntoResponse {
-    let store = state.contacts.read().await;
-    let revocations: Vec<serde_json::Value> = store
-        .revocations()
-        .iter()
-        .map(|r| {
-            serde_json::json!({
-                "agent_id": hex::encode(r.agent_id.0),
-                "reason": r.reason,
-                "timestamp": r.timestamp,
-                "revoker_id": r.revoker_id.map(|id| hex::encode(id.0))
-            })
-        })
-        .collect();
-
-    (
-        StatusCode::OK,
-        Json(serde_json::json!({ "ok": true, "revocations": revocations })),
-    )
-}
-
-/// POST /contacts/:agent_id/machines/:machine_id/pin — pin a machine for identity verification.
-async fn pin_machine(
-    State(state): State<Arc<AppState>>,
-    Path((agent_id_hex, machine_id_hex)): Path<(String, String)>,
-) -> impl IntoResponse {
-    let agent_id = match parse_agent_id_hex(&agent_id_hex) {
-        Ok(id) => id,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(serde_json::json!({ "ok": false, "error": e })),
-            );
-        }
-    };
-
-    let machine_bytes = match hex::decode(&machine_id_hex) {
-        Ok(bytes) if bytes.len() == 32 => {
-            let mut arr = [0u8; 32];
-            arr.copy_from_slice(&bytes);
-            arr
-        }
-        _ => {
-            return bad_request("invalid machine_id hex");
-        }
-    };
-    let machine_id = MachineId(machine_bytes);
-
-    let mut store = state.contacts.write().await;
-    let pinned = store.pin_machine(&agent_id, &machine_id);
-
-    (
-        StatusCode::OK,
-        Json(serde_json::json!({ "ok": true, "pinned": pinned })),
-    )
-}
-
-/// DELETE /contacts/:agent_id/machines/:machine_id/pin — unpin a machine.
-async fn unpin_machine(
-    State(state): State<Arc<AppState>>,
-    Path((agent_id_hex, machine_id_hex)): Path<(String, String)>,
-) -> impl IntoResponse {
-    let agent_id = match parse_agent_id_hex(&agent_id_hex) {
-        Ok(id) => id,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(serde_json::json!({ "ok": false, "error": e })),
-            );
-        }
-    };
-
-    let machine_bytes = match hex::decode(&machine_id_hex) {
-        Ok(bytes) if bytes.len() == 32 => {
-            let mut arr = [0u8; 32];
-            arr.copy_from_slice(&bytes);
-            arr
-        }
-        _ => {
-            return bad_request("invalid machine_id hex");
-        }
-    };
-    let machine_id = MachineId(machine_bytes);
-
-    let mut store = state.contacts.write().await;
-    let unpinned = store.unpin_machine(&agent_id, &machine_id);
-
-    (
-        StatusCode::OK,
-        Json(serde_json::json!({ "ok": true, "unpinned": unpinned })),
-    )
-}
 
 /// POST /trust/evaluate — evaluate trust decision for an (agent, machine) pair.
 async fn evaluate_trust(
